start clusterappman with api-server
