+++ cmd/kore-apiserver/server.go
@@ -72,6 +72,7 @@ func invoke(ctx *cli.Context) error {
 			CertificateAuthority:       ctx.String("certificate-authority"),
 			CertificateAuthorityKey:    ctx.String("certificate-authority-key"),
 			ClusterAppManImage:         ctx.String("clusterappman-image"),
+			ManagedDependencies:        ctx.Bool("enable-managed-dependencies"),
 			EnableClusterProviderCheck: ctx.Bool("enable-cluster-provider-check"),
 			FeatureGates:               featuresGates,
 			HMAC:                       ctx.String("kore-hmac"),
