+++ pkg/controllers/management/kubernetes/helper.go
@@ -20,23 +20,15 @@ import (
 	"bytes"
 	"context"
 	"errors"
-	"fmt"
 	"text/template"
-	"time"
 
 	clustersv1 "github.com/appvia/kore/pkg/apis/clusters/v1"
-	clusterappman "github.com/appvia/kore/pkg/clusterappman"
-	"github.com/appvia/kore/pkg/kore"
 	"github.com/appvia/kore/pkg/utils/kubernetes"
 	log "github.com/sirupsen/logrus"
 
 	"github.com/Masterminds/sprig"
 	yaml "github.com/ghodss/yaml"
 	apps "k8s.io/api/apps/v1"
-	appsv1 "k8s.io/api/apps/v1"
-	core "k8s.io/api/core/v1"
-	v1 "k8s.io/api/core/v1"
-	rbacv1 "k8s.io/api/rbac/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -104,140 +96,12 @@ func ListAllClustersInTeam(ctx context.Context, cc client.Client, namespace stri
 
 }
 
-// WaitOnStatus will wait until the status object exists
-// TODO: define a status object suitabvle for overall cluster status
-func WaitOnStatus(ctx context.Context, cc client.Client) error {
-	// WaitOnStatus checks the status of the job and if not successful returns the error
-	for {
-		select {
-		case <-ctx.Done():
-			return errors.New("context has been cancelled")
-		default:
-		}
-
-		err := func() error {
-			exists, err := StatusExists(ctx, cc)
-			if err != nil {
-				return err
-			}
-			if exists {
-				return nil
-			}
-
-			return errors.New("Kore cluster manager has not reported status yet")
-		}()
-		if err == nil {
-			return nil
-		}
-		time.Sleep(10 * time.Second)
-	}
-}
-
-// StatusExists checks if the status exists already
-func StatusExists(ctx context.Context, cc client.Client) (bool, error) {
-	return HasConfigMap(ctx, cc, clusterappman.StatusCongigMap)
-}
-
-// CreateConfig creates a configmap for configuring the kore cluster manager
-func CreateConfig(ctx context.Context, cc client.Client, params Parameters) error {
-	b, err := yaml.Marshal(params)
-	if err != nil {
-		return errors.New("can not marshall params into yaml")
-	}
-	// Specify the parameters in the config map
-	cm := &core.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      clusterappmanConfig,
-			Namespace: clusterappmanNamespace,
-		},
-		Data: (map[string]string{"clusterconfig": string(b)}),
-	}
-	if _, err := kubernetes.CreateOrUpdate(ctx, cc, cm); err != nil {
-		return err
-	}
-	return nil
-}
-
-// GetConfig will retrieve the cluster configuration values
-func GetConfig(ctx context.Context, cc client.Client) (params *Parameters, err error) {
-	p := &Parameters{}
-	cm := &v1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      clusterappmanConfig,
-			Namespace: clusterappmanNamespace,
-		},
-	}
-	exists, err := kubernetes.GetIfExists(ctx, cc, cm)
-	if err != nil {
-		return p, fmt.Errorf(
-			"error obtaining config for %s/%s - %s",
-			clusterappmanNamespace,
-			clusterappmanConfig,
-			err,
-		)
-	}
-	if exists {
-		// get the data from the configmap data key
-		b := []byte(cm.Data[clusterappmanConfigKey])
-		if err := yaml.Unmarshal(b, p); err != nil {
-			return p, fmt.Errorf(
-				"unable to deserialize yaml data from config map: %s/%s, key: %s",
-				clusterappmanNamespace,
-				clusterappmanConfig,
-				clusterappmanConfigKey,
-			)
-		}
-		return p, nil
-	}
-	return p, fmt.Errorf(
-		"missing configmap %s in namespace %s",
-		clusterappmanConfig,
-		clusterappmanNamespace,
-	)
-}
-
-// ConfigExists check if the cluster configuration exists
-func ConfigExists(ctx context.Context, cc client.Client) (bool, error) {
-	return HasConfigMap(ctx, cc, clusterappmanConfig)
-}
-
-// HasConfigMap checks if a configmap exists
-func HasConfigMap(ctx context.Context, cc client.Client, name string) (bool, error) {
-	return kubernetes.CheckIfExists(ctx, cc, &v1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      name,
-			Namespace: clusterappmanNamespace,
-		},
-	})
-}
-
-// NamespaceExists checks if the bootstrap job there
-func NamespaceExists(ctx context.Context, cc client.Client) (bool, error) {
-	return kubernetes.CheckIfExists(ctx, cc, &core.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: clusterappmanNamespace,
-		},
-	})
-}
-
 // DeploymentExists checks if the bootstrap job there
 func DeploymentExists(ctx context.Context, cc client.Client, name, namespace string) (bool, error) {
 	return kubernetes.CheckIfExists(ctx, cc, &apps.Deployment{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      name,
 			Namespace: namespace,
-		},
-	})
-}
-
-// EnsureNamespace creates a namespace for the clusterappmanager if required
-func EnsureNamespace(ctx context.Context, cc client.Client, namespace string) error {
-	return kubernetes.EnsureNamespace(ctx, cc, &core.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: namespace,
-			Labels: map[string]string{
-				kore.Label("owned"): "true",
-			},
 		},
 	})
 }
@@ -260,82 +124,6 @@ func GetDeployment(ctx context.Context, cc client.Client, name, namespace string
 	}
 
 	return deployment, nil
-}
-
-// CreateOrUpdateClusterAppManDeployment will reconcile the clusterappman deployment
-func CreateOrUpdateClusterAppManDeployment(ctx context.Context, cc client.Client, image string) error {
-	name := clusterappmanDeployment
-	if _, err := kubernetes.CreateOrUpdate(ctx, cc, &appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      clusterappmanDeployment,
-			Namespace: clusterappmanNamespace,
-			Labels: map[string]string{
-				"name": name,
-			},
-		},
-		Spec: appsv1.DeploymentSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: map[string]string{
-					"name": name,
-				},
-			},
-			Strategy: appsv1.DeploymentStrategy{
-				Type: appsv1.RollingUpdateDeploymentStrategyType,
-			},
-			Template: v1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{
-						"name": name,
-					},
-				},
-				Spec: v1.PodSpec{
-					ServiceAccountName: "clusterappman",
-					Containers: []v1.Container{
-						{
-							Name:  name,
-							Image: image,
-							Env: []v1.EnvVar{
-								{
-									Name:  "IN_CLUSTER",
-									Value: "true",
-								},
-							},
-							Command: []string{
-								"/bin/kore-clusterappman",
-							},
-						},
-					},
-				},
-			},
-		},
-	}); err != nil {
-		return err
-	}
-	return nil
-}
-
-// CreateClusterManClusterRoleBinding creates (or updates) the cluster role binding required for the clusterappman
-func CreateClusterManClusterRoleBinding(ctx context.Context, cc client.Client) error {
-	if _, err := kubernetes.CreateOrUpdateManagedClusterRoleBinding(ctx, cc, &rbacv1.ClusterRoleBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "kore:clusterappman",
-		},
-		Subjects: []rbacv1.Subject{
-			{
-				Kind:      "ServiceAccount",
-				Name:      "clusterappman",
-				Namespace: clusterappman.KoreNamespace,
-			},
-		},
-		RoleRef: rbacv1.RoleRef{
-			APIGroup: "rbac.authorization.k8s.io",
-			Kind:     "ClusterRole",
-			Name:     "cluster-admin",
-		},
-	}); err != nil {
-		return fmt.Errorf("error tying to apply kore clusterappman clusterrole %q", err)
-	}
-	return nil
 }
 
 // DecodeInTo decodes the template into the thing
