+++ pkg/kore/types.go
@@ -123,6 +123,8 @@ type Config struct {
 	CertificateAuthorityKey string `json:"certificate-authority-key,omitempty"`
 	// ClusterAppManImage is the image to use for cluster application management
 	ClusterAppManImage string `json:"cluster-app-man-image,omitempty"`
+	// ManagedDependencies will run clusterappman in the apiserver configured
+	ManagedDependencies bool `json:"enable-managed-dependencies"`
 	// DEX is the config required to configure dex
 	DEX DEX `json:"dex,omitempty"`
 	// EnableClusterProviderCheck indicate the k8s controller should check the status of the
